feat: optimize useRequest
