+++ scaffolds/ice-design-pro/src/stores/userProfile.js
@@ -1,4 +1,4 @@
-import request from '@/utils/request';
+import { request } from '@/utils/request';
 import { userProfile } from '@/config/dataSource';
 
 export default {
